requirements.txt
pycryptodome==3.11.0
typer==0.4.0